* src/libcommon/str.c, src/libcommon/str.h (strcatf):   Fixed bug (found by coverity) where strcatf()'s size_t return type   prevented the error code of -1 from being returned to the caller. * src/munged/crypto.c (openssl_log_msg):   Specified void cast to discard return value from strcatf().

+++ src/libcommon/str.h
@@ -42,7 +42,7 @@ char * strdupf (const char *fmt, ...);
  *  The caller is responsible for freeing this new string.
  */
 
-size_t strcatf (char *dst, size_t size, const char *fmt, ...);
+int strcatf (char *dst, size_t size, const char *fmt, ...);
 /*
  *  Concatenates the string specified by the format-string [fmt] to
  *    the NUL-terminated string [dst] within a buffer of size [size].
